Initial commit
This code implements enough to allow an object model to be defined using
 + define-attribute-type
 + define-property-type
 + define-property-constructor
 + define-object-type
 + define-object-constructor

The object model can then be used to make, create, read, save! and
manipulate instances of those objects.

Objects are persisted to an SQLite database.

A few initial attribute types are provided:
 + integer
 + text
 + blob
 + boolean
 + symbol
 + wallclock
 + node-id/current-node
 + timestamp/next
 + current-user/node-id
 + current-user/seqno

initialise-persistent-counter and load-persistent-counter provide a very
simple consistency model. This just stores the latest counter value in a
file every time a counter is consumed. This means that data models created
with this can only be used in one process at a time. This restriction will
be lifted in future work.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>